requirements_api.txt
amqp==                     5.1.1
anyio==                    3.6.2
asgiref==                3.6.0
astor==                  0.8.1
astunparse==               1.6.3
attrdict ==                2.0.1
Babel==                    2.12.1
bce-python-sdk==           0.8.87
beautifulsoup4==           4.12.2
bidict==                   0.22.1
billiard==                 3.6.4.0
cached-property==          1.5.2
cachetools==               4.2.4
celery==                   5.2.7
certifi==                  2022.12.7
cfgv==                     3.3.1
chardet==                  3.0.4
charset-normalizer==       3.1.0
click==                    8.1.3
click-didyoumean==         0.3.0
click-plugins==            1.1.1
click-repl==               0.2.0
colorama==                 0.4.4
craft-text-detector==      0.1.8
cssselect==                1.2.0
cssutils==                 2.7.1
cycler==                   0.11.0
Cython==                   3.0.0
cython-bbox==              0.1.3
decorator==                5.1.1
distlib==                  0.3.6
distro==                   1.5.0
Django==                   3.2.18
djangorestframework==      3.14.0
docker==                   4.3.1
easydict==                 1.10
editdistance==             0.6.2
efficientnet==             1.0.0
essential-generators==     1.0
et-xmlfile==               1.1.0
filelock==                 3.10.7
fire==                     0.5.0
Flask==                    2.2.5
flask-babel==              3.1.0
Flask-SocketIO==           5.1.0
flatbuffers==              1.12
fonttools==                4.38.0
fsspec==                   2023.1.0
future==                   0.18.3
gast==                     0.4.0
gdown==                   4.7.1
google-auth==              1.35.0
google-auth-oauthlib==     0.4.6
google-pasta==             0.2.0
GPUtil==                   1.4.0
grpcio==                   1.34.1
h11==                      0.12.0
h5py==                     3.1.0
httpcore==                 0.13.7
httpx==                    0.18.0
huggingface-hub==          0.14.1
identify==                 2.5.22
idna==                     2.10
imageio==                  2.28.1
imagesize==                1.2.0
imgaug==                   0.4.0
importlib-metadata==       6.1.0
itsdangerous==             2.1.2
Jinja2==                   3.1.2
Keras-Applications==       1.0.8
keras-nightly==            2.5.0.dev2021032900
keras-ocr==                0.8.9
Keras-Preprocessing==      1.1.2
kiwisolver==               1.4.4
kombu==                    5.2.4
lanms-neo==                1.0.2
lap==                      0.4.0
lmdb==                     1.4.1
lxml==                     4.9.2
Markdown==                 3.4.3
MarkupSafe==               2.1.2
matplotlib==               3.5.3
nest-asyncio==             1.5.6
networkx==                 2.6.3
nodeenv==                  1.7.0
numpy==                    1.19.5
nvidia-cublas-cu11==       11.10.3.66
nvidia-cuda-nvrtc-cu11==   11.7.99
nvidia-cuda-runtime-cu11== 11.7.99
nvidia-cudnn-cu11==        8.5.0.96
oauthlib==                 3.2.2
opencv-contrib-python==    4.6.0.66
opencv-python==            4.6.0.66
openpyxl==                 3.1.2
opt-einsum==               3.3.0
packaging==                23.0
paddle-bfloat==            0.1.7
paddleocr==                2.6.1.1
paddlepaddle==             2.4.2
pandas==                   1.3.5
pdf2docx==                 0.5.6
Pillow==                   8.3.2
pip==                      23.2
platformdirs==             3.2.0
Polygon3==                 3.0.9.1
pre-commit==               2.21.0
premailer==                3.10.0
prompt-toolkit==          3.0.38
protobuf==                 3.20.0
psutil==                   5.8.0
psycopg2-binary==          2.9.6
py-cpuinfo==               8.0.0
pyasn1==                   0.4.8
pyasn1-modules==           0.2.8
pyclipper==                1.3.0.post4
pycryptodome==             3.18.0
PyMuPDF==                  1.20.2
pyparsing==                3.0.9
PySocks==                  1.7.1
python-dateutil==          2.8.2
python-docx==              0.8.11
python-engineio==          4.4.1
python-socketio==          5.3.0
pytz==                     2023.3
PyWavelets==               1.3.0
PyYAML==                   5.4.1
rapidfuzz==                3.2.0
rarfile==                  4.0
regex==                    2023.5.5
requests==                 2.25.0
requests-oauthlib==        1.3.1
rfc3986==                  1.5.0
rsa==                      4.9
scikit-image==             0.19.3
scipy==                    1.4.1
seaborn==                  0.12.2
sentencepiece==            0.1.97
setuptools==               65.6.3
Shapely==                  1.7.0
shutup==                   0.2.0
six==                      1.15.0
sniffio==                  1.3.0
soupsieve==                2.4.1
SQLAlchemy==               1.3.22
sqlparse==                0.4.3
tensorboard==              2.11.2
tensorboard-data-server==  0.6.1
tensorboard-plugin-wit==   1.8.1
tensorflow-estimator==     2.5.0
tensorflow-gpu==           2.5.0
termcolor==                1.1.0
theos-ai==                 0.0.28
tifffile==                 2021.11.2
tokenizers==               0.12.1
torch==                    1.13.1
torchvision==              0.14.1
tqdm==                     4.48.0
transformers==             4.22.1
typing-extensions==        3.7.4.3
urllib3==                  1.26.7
validators==               0.20.0
vine==                     5.0.0
virtualenv==               20.21.0
visualdl==                 2.5.3
wcwidth==                  0.2.6
websocket-client==         1.5.1
websockets==               10.3
Werkzeug==                 2.2.3
wheel==                    0.38.4
wrapt==                    1.12.1
yaspin==                   1.1.0
zipp==                     3.15.0
